requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-environ==0.8.1
django-filter==22.1
django-rest-auth==0.9.5
django-simple-history==3.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
